Cleaned up Sidebar code a little
