+++ www/src/templates/blog.js
@@ -11,11 +11,11 @@
 
 'use strict';
 
-import MarkdownPage from 'components/MarkdownPage';
 import PropTypes from 'prop-types';
 import React from 'react';
+import MarkdownPage from 'components/MarkdownPage';
+import {createLinkBlog} from 'utils/createLink';
 
-// TODO This is hacky
 const toSectionList = allMarkdownRemark => [
   {
     title: 'Recent Posts',
@@ -34,6 +34,7 @@ const toSectionList = allMarkdownRemark => [
 const Blog = ({data, location}) => (
   <MarkdownPage
     authors={data.markdownRemark.frontmatter.author}
+    createLink={createLinkBlog}
     date={data.markdownRemark.fields.date}
     location={location}
     ogDescription={data.markdownRemark.excerpt}
